Created basic jest, supertest api test just to check if 200 response
