+++ package.json
@@ -6,13 +6,14 @@
   "scripts": {
     "nodemon": "nodemon --watch tsDist ",
     "server": "nodemon --watch tsDist tsDist/server/server",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
     "@types/express": "^4.16.0",
+    "@types/jest": "^23.3.1",
     "@types/lokijs": "^1.5.2",
     "@types/node": "^10.9.4",
     "@types/supertest": "^2.0.5",
@@ -24,6 +25,7 @@
     "nodemon": "^1.18.4",
     "prettier": "^1.14.2",
     "supertest": "^3.2.0",
+    "ts-jest": "^23.1.4",
     "tslint": "^5.11.0",
     "tslint-config-prettier": "^1.15.0",
     "tslint-plugin-prettier": "^1.3.0",
