+++ src/server/routes/pollRouter.ts
@@ -4,7 +4,8 @@ import pollsModel from "../models/pollsModel";
 const pollRouter = express.Router();
 
 pollRouter.route("/").get((req, res) => {
-  const poll = pollsModel.findOne({ name: "test2" });
-  console.log(poll);
+  const polls: object[] = pollsModel.find();
+  console.log("/api/polls was accessed");
+  res.json(polls);
 });
 export default pollRouter;
