+++ tsDist/server/app.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const bodyParser = require("body-parser");
+const express = require("express");
+const pollRouter_1 = require("./routes/pollRouter");
+const app = express();
+app.use(bodyParser.urlencoded({ extended: true }));
+app.use(bodyParser.json());
+app.use("/api/polls", pollRouter_1.default);
+exports.default = app;
