+++ tsDist/server/server.js
@@ -1,11 +1,5 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const bodyParser = require("body-parser");
-const express = require("express");
-const pollRouter_1 = require("./routes/pollRouter");
-const app = express();
+const app_1 = require("./app");
 const port = process.env.PORT || 8000;
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
-app.use("/api/polls", pollRouter_1.default);
-app.listen(port, () => console.log(`App is running on port: ${port}`));
+app_1.default.listen(port, () => console.log(`App is running on port: ${port}`));
